A generic command dispatcher using asyncio.
This is my first attempt to use asyncio other than webservices. My
motivation was to create a helper that could launch or execute programs
for me in the background, similar to `GNU parallel` but written in
python, and a test concept for if this can be achieved with asyncio. If
so, what problems should I encounter along the way and how to overcome
those, if any.